% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbouSummarise.R
\name{bbouMakeFigures}
\alias{bbouMakeFigures}
\title{Make figures based on fitted survival and recruitment models}
\usage{
bbouMakeFigures(
  surv_fit,
  recruit_fit,
  fig_dir,
  i18n = NULL,
  ht = 400,
  wt = 600,
  show_interpolated = TRUE
)
}
\arguments{
\item{surv_fit}{result of \code{bboutools::bb_fit_survival}}

\item{recruit_fit}{result of \code{bboutools::bb_fit_recruitment}}

\item{fig_dir}{file path. Directory where figures should be saved}

\item{i18n}{shiny.i18n object. Used of translation in Shiny app.}

\item{ht}{numeric. Height in pixels}

\item{wt}{numeric. Width in pixels}
}
\value{
Nothing. Figures are saved to \code{fig_dir}.
}
\description{
Make figures based on fitted survival and recruitment models
}
