% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbouSummarise.R
\name{bbouMakeSummaryTable}
\alias{bbouMakeSummaryTable}
\title{Create summary table of demographic rates from survival and recruitment surveys}
\usage{
bbouMakeSummaryTable(
  surv_data,
  recruit_data,
  N0,
  shiny_progress = FALSE,
  return_mcmc = FALSE,
  i18n = NULL,
  ...
)
}
\arguments{
\item{surv_data}{dataframe. Survival data in bboudata format}

\item{recruit_data}{dataframe. Recruitment data in bboudata format}

\item{N0}{dataframe. Initial population estimates, required columns are PopulationName and N0}

\item{return_mcmc}{boolean. If TRUE return fitted survival and recruitment models. Default FALSE.}

\item{...}{}
}
\description{
Create summary table of demographic rates from survival and recruitment surveys
}
\examples{
s_data <- rbind(bboudata::bbousurv_a, bboudata::bbousurv_b)
r_data <- rbind(bboudata::bbourecruit_a, bboudata::bbourecruit_b)
fitDemographicRates(s_data, r_data, 500, FALSE)
}
