% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_caribou_demog_app}
\alias{run_caribou_demog_app}
\title{Run the Boreal Caribou Demographic Projection Explorer}
\usage{
run_caribou_demog_app(private = FALSE, lang = "en", allow_update_data = TRUE)
}
\arguments{
\item{private}{If private then you will need to follow instructions in console to authenticate google drive}
}
\value{
launches the app in your default browser
}
\description{
Run the Boreal Caribou Demographic Projection Explorer
}
