% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaribouDemographyBasicApp-package.R
\docType{package}
\name{CaribouDemographyBasicApp-package}
\alias{CaribouDemographyBasicApp}
\alias{CaribouDemographyBasicApp-package}
\title{CaribouDemographyBasicApp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LandSciTech/CaribouDemographyBasicApp}
  \item Report bugs at \url{https://github.com/LandSciTech/CaribouDemographyBasicApp/issues}
}

}
\author{
\strong{Maintainer}: Sarah Endicott \email{sarah.endicott@canada.ca} (\href{https://orcid.org/0000-0001-9644-5343}{ORCID})

}
\keyword{internal}
