% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaribouDemographyBasicApp-package.R
\docType{package}
\name{CaribouDemographyBasicApp-package}
\alias{CaribouDemographyBasicApp}
\alias{CaribouDemographyBasicApp-package}
\title{CaribouDemographyBasicApp: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
