% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Get updated data from google spreadsheet}
\usage{
update_data(
  survey_url,
  inst_dir = system.file(package = "CaribouDemographyBasicApp"),
  i18n = NULL,
  lang = "en",
  shiny_progress = FALSE
)
}
\arguments{
\item{survey_url}{sheet url}

\item{inst_dir}{dir where results are saved}

\item{shiny_progress}{Is this inside a shiny app and called with \code{shiny::withProgress}}
}
\description{
Get updated data from google spreadsheet
}
