% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Get updated data from google spreadsheet}
\usage{
update_data(
  survey_url,
  save_dir = tools::R_user_dir("CaribouDemographyBasicApp", "data"),
  i18n = NULL,
  lang = "en",
  shiny_progress = FALSE
)
}
\arguments{
\item{survey_url}{sheet url}

\item{save_dir}{dir where results are saved}

\item{i18n}{shiny.i18n translator}

\item{lang}{language to select in translator}

\item{shiny_progress}{Is this inside a shiny app and called with \code{shiny::withProgress}}
}
\value{
A data frame with the population estimates.
}
\description{
Get updated data from google spreadsheet
}
\examples{
googlesheets4::gs4_deauth()
update_data("https://docs.google.com/spreadsheets/d/1i53nQrJXgrq3B6jO0ATHhSIbibtLq5TmmFL-PxGQNm8/edit?usp=sharing",
            save_dir = tempdir())
}
